Update install-panel.sh
I was experimenting with Zorin as that's all I had laying around on a USB Stick at the time, as the OS is built on top of Ubuntu 18 there isn't much difference in the OS other than the overhauled UI. So I changed your script to include it locally. Just suggesting the change in case anyone else uses this. I was able to install and use it with those changes.