Package not installed when function is ran
The package software-properties-common isn't installed before the command it provides (add-apt-repository) is ran.
The fix basically installs the necessary package. Though it is suggestion that this package is installed another way, or since the command 'apt install software-properties-common -y' is already within the script, that command should be ran before this script.

Though for now, this fixes the error message the start of running the script saying "Error: add-apt-repository. command not found"