Just tried to make the password section better.
There were issues with the previous one where, if you copy and paste a large password, sometimes it doesn't display the * and displayers the text.
When I first had this come up I didn't want to continue using a stronger password, because I was afriad I would know the password after, since it would be inputted incorrectly.

This follows the traditional style of Linux passwords, when you put in the password, no text comes up when you type, and it asks you to confirm the password with no text still coming up. This should still with with the multiple variables in the script, you can use it by doing `password_input VARIABLE` then the VARIABLE would be the password that is entered.

Haven't tested it using the main script, only tested it sectioned off, since I think its a pretty small change.