panel: handle exit status 1 in ask_assume_ssl
When executed, the exit status of a function is the exit status of the last command executed in the body. Since the expression would return exit 1, the function ask_assume_ssl would cause the script to exit if the user selected not to ASSUME_SSL (the expression is the last command of the function).