Completely rewritten
I decided to completely rewrite the script, as I realized that copying and pasting will only create problems for me.
Note: I made the script without choosing SSL Certificate and more specific settings because I do not have much time, if anyone wants to help with this, will be welcome.